requirements.txt
# Core ML Libraries
torch>=2.5.1
torchvision>=0.20.0
transformers>=4.56.2
diffusers>=0.30.0
accelerate>=0.33.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0
segment-anything>=1.0

# GroundingDINO
groundingdino-py>=0.4.0

# LaMa Inpainting
lama-cleaner>=1.2.0

# ControlNet
controlnet-aux>=0.4.0

# Background Removal
rembg>=2.0.50

# Depth Estimation
zoedepth>=0.1.0

# Upscaling
basicsr>=1.4.2
realesrgan>=0.3.0

# SVG Processing
svgpathtools>=1.4.4
shapely>=2.0.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Error Handling & Retry
tenacity>=8.2.3

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# CLI
click>=8.1.7

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
slowapi>=0.1.9
jinja2>=3.1.2

# Color Science (optional, for exact CIEDE2000)
colorspacious>=2.0.0

# GPU-accelerated computer vision
kornia>=0.7.0

# LoRA Training
peft>=0.13.0
albumentations>=1.3.0

# Memory Optimization (optional but recommended)
xformers>=0.0.23
bitsandbytes>=0.43.0

# Monitoring (optional)
prometheus-client>=0.18.0

# HTTP requests (for model downloads)
requests>=2.31.0

